4.21

+++ part4/controllers/blogs.js
@@ -35,8 +35,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, config.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === decodedToken.id) {
+    console.log('Correct user, deleting');
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    console.log('Denied: wrong user');
+    response.status(401).json({ error: 'Only own blog entries can be deleted' });
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
